MDL-79737 theme_boost: issues when modal toggles due to window resize
When you open a drawer while the page is below the "medium" width
cutoff, the drawer opens as a modal. The underlying page gets grayed out
and goes no-scrollable. When you close the modal drawer, these steps are
reversed and the page becomes visible and scrollable again.

However, when you open a modal drawer and then resize the window to be
wider than the cutoff, the drawer will change from modal to non-modal
and the page becomes visible, but the scroll disabling is *not* removed
as it should be. Similarly, if you open a drawer while the page is wide,
and then narrow the page beyond the cutoff, the drawer will change to a
modal and the page will be grayed out, but scrolling on the underlying
page is *not* disabled as it should be.

This patch corrects these issues so that scrolling of the underlying
page is disabled/enabled when the drawer goes in and out of modal form
due to window resizes, the same as when the modal drawer is opened and
closed via mouse clicks.